refactor(86byb8er1): renamed configurations folder and removed jwt key
