+++ src/main/java/pl/wip/wipplatformbe/config/ApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package pl.wip.wipplatformbe.configurations;
+package pl.wip.wipplatformbe.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
