+++ src/main/java/pl/wip/wipplatformbe/config/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package pl.wip.wipplatformbe.configurations;
+package pl.wip.wipplatformbe.config;
 
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.context.annotation.Bean;
