+++ src/main/resources/application.properties
@@ -1,11 +1,7 @@
-# When running in a container
-#spring.data.mongodb.host=mongodb
-# When running locally
-#spring.data.mongodb.host=localhost
+# When running locally comment line below
+spring.data.mongodb.host=mongodb
 spring.data.mongodb.port=27017
 spring.data.mongodb.database=wip
-#spring.data.mongodb.username=${MONGODB_USERNAME}
-#spring.data.mongodb.password=${MONGODB_PASSWORD}
 
-security.jwt.secret-key=3cfa76ef14937c1c0ea519f8fc057a80fcd04a7420f8e8bcd0a7567c272e007b
+security.jwt.secret-key=${JWT_SECRET_KEY}
 security.jwt.expiration-time=3600000
